Update TOF_Operational_V0.ino
Tested TOF device with balloon, works great. 
Adding sleep and wake up routines, included a general lowpass filter for the pressure sensor (I will tune this in the future) and set up some indicators for when the device is going into sleep or waking up.
Will spend time tomorrow trying to get the system to time itself to know when to go to sleep and when not to.